+++ docs/source/history.md
@@ -92,6 +92,8 @@ Enjoy - we are awaiting comments and issues, thanks for testing.
 
 ## v0.2.3-alpha
 
+- Input states
+  - Improved: `SingleStateStates` now supports profiles
 - Turbine models
   - New model `PowerMask`: Derates or boosts a turbine
 - Examples